feat: implement profile analytics view with integrated recommendation system and database support
